update all testcase
