+++ src/test/java/org/jsoup/nodes/CommentTest.java
@@ -1,4 +1,4 @@
-package testoracle;
+package org.jsoup.nodes;
 
 import static org.junit.Assert.*;
 
@@ -15,31 +15,5 @@ public class CommentTest {
 		assertEquals("\n<!--This is wjdebug comment-->", c.toString());
 	}
 
-	/*
-	@Test
-	public void testNodeName() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testOuterHtmlHead() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testOuterHtmlTail() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testToString() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testGetData() {
-		fail("Not yet implemented");
-	}
-	*/
 
 }
