+++ src/test/java/org/jsoup/nodes/FormElementTest.java
@@ -12,136 +12,174 @@ import static org.junit.Assert.*;
  * Tests for FormElement
  *
  * @author Jonathan Hedley
+ * @Contributor : Heekyo Lee (leeheekyo)
  */
 public class FormElementTest {
-    @Test public void hasAssociatedControls() {
-        //"button", "fieldset", "input", "keygen", "object", "output", "select", "textarea"
-        String html = "<form id=1><button id=1><fieldset id=2 /><input id=3><keygen id=4><object id=5><output id=6>" +
-                "<select id=7><option></select><textarea id=8><p id=9>";
-        Document doc = Jsoup.parse(html);
+	 @Test public void hasAssociatedControls() {
+	        //"button", "fieldset", "input", "keygen", "object", "output", "select", "textarea"
+	        String html = "<form id=1><button id=1><fieldset id=2 /><input id=3><keygen id=4><object id=5><output id=6>" +
+	                "<select id=7><option></select><textarea id=8><p id=9>";
+	        Document doc = Jsoup.parse(html);
 
-        FormElement form = (FormElement) doc.select("form").first();
-        assertEquals(8, form.elements().size());
-    }
+	        FormElement form = (FormElement) doc.select("form").first();
+	        assertEquals(8, form.elements().size());
+	    }
 
-    @Test public void createsFormData() {
-        String html = "<form><input name='one' value='two'><select name='three'><option value='not'>" +
-                "<option value='four' selected><option value='five' selected><textarea name=six>seven</textarea>" +
-                "<input name='seven' type='radio' value='on' checked><input name='seven' type='radio' value='off'>" +
-                "<input name='eight' type='checkbox' checked><input name='nine' type='checkbox' value='unset'>" +
-                "<input name='ten' value='text' disabled>" +
-                "</form>";
-        Document doc = Jsoup.parse(html);
-        FormElement form = (FormElement) doc.select("form").first();
-        List<Connection.KeyVal> data = form.formData();
+	    @Test public void createsFormData() {
+	        String html = "<form><input name='one' value='two'><select name='three'><option value='not'>" +
+	                "<option value='four' selected><option value='five' selected><textarea name=six>seven</textarea>" +
+	                "<input name='seven' type='radio' value='on' checked><input name='seven' type='radio' value='off'>" +
+	                "<input name='eight' type='checkbox' checked><input name='nine' type='checkbox' value='unset'>" +
+	                "<input name='ten' value='text' disabled>" +
+	                "</form>";
+	        Document doc = Jsoup.parse(html);
+	        FormElement form = (FormElement) doc.select("form").first();
+	        List<Connection.KeyVal> data = form.formData();
 
-        assertEquals(6, data.size());
-        assertEquals("one=two", data.get(0).toString());
-        assertEquals("three=four", data.get(1).toString());
-        assertEquals("three=five", data.get(2).toString());
-        assertEquals("six=seven", data.get(3).toString());
-        assertEquals("seven=on", data.get(4).toString()); // set
-        assertEquals("eight=on", data.get(5).toString()); // default
-        // nine should not appear, not checked checkbox
-        // ten should not appear, disabled
-    }
+	        assertEquals(6, data.size());
+	        assertEquals("one=two", data.get(0).toString());
+	        assertEquals("three=four", data.get(1).toString());
+	        assertEquals("three=five", data.get(2).toString());
+	        assertEquals("six=seven", data.get(3).toString());
+	        assertEquals("seven=on", data.get(4).toString()); // set
+	        assertEquals("eight=on", data.get(5).toString()); // default
+	        // nine should not appear, not checked checkbox
+	        // ten should not appear, disabled
+	    }
 
-    @Test public void createsSubmitableConnection() {
-        String html = "<form action='/search'><input name='q'></form>";
-        Document doc = Jsoup.parse(html, "http://example.com/");
-        doc.select("[name=q]").attr("value", "jsoup");
+	    @Test public void createsSubmitableConnection() {
+	        String html = "<form action='/search'><input name='q'></form>";
+	        Document doc = Jsoup.parse(html, "http://example.com/");
+	        doc.select("[name=q]").attr("value", "jsoup");
 
-        FormElement form = ((FormElement) doc.select("form").first());
-        Connection con = form.submit();
+	        FormElement form = ((FormElement) doc.select("form").first());
+	        Connection con = form.submit();
 
-        assertEquals(Connection.Method.GET, con.request().method());
-        assertEquals("http://example.com/search", con.request().url().toExternalForm());
-        List<Connection.KeyVal> dataList = (List<Connection.KeyVal>) con.request().data();
-        assertEquals("q=jsoup", dataList.get(0).toString());
+	        assertEquals(Connection.Method.GET, con.request().method());
+	        assertEquals("http://example.com/search", con.request().url().toExternalForm());
+	        List<Connection.KeyVal> dataList = (List<Connection.KeyVal>) con.request().data();
+	        assertEquals("q=jsoup", dataList.get(0).toString());
 
-        doc.select("form").attr("method", "post");
-        Connection con2 = form.submit();
-        assertEquals(Connection.Method.POST, con2.request().method());
-    }
+	        doc.select("form").attr("method", "post");
+	        Connection con2 = form.submit();
+	        assertEquals(Connection.Method.POST, con2.request().method());
+	    }
 
-    @Test public void actionWithNoValue() {
-        String html = "<form><input name='q'></form>";
-        Document doc = Jsoup.parse(html, "http://example.com/");
-        FormElement form = ((FormElement) doc.select("form").first());
-        Connection con = form.submit();
+	    @Test public void actionWithNoValue() {
+	        String html = "<form><input name='q'></form>";
+	        Document doc = Jsoup.parse(html, "http://example.com/");
+	        FormElement form = ((FormElement) doc.select("form").first());
+	        Connection con = form.submit();
 
-        assertEquals("http://example.com/", con.request().url().toExternalForm());
-    }
+	        assertEquals("http://example.com/", con.request().url().toExternalForm());
+	    }
 
-    @Test public void actionWithNoBaseUri() {
-        String html = "<form><input name='q'></form>";
-        Document doc = Jsoup.parse(html);
-        FormElement form = ((FormElement) doc.select("form").first());
+	    @Test public void actionWithNoBaseUri() {
+	        String html = "<form><input name='q'></form>";
+	        Document doc = Jsoup.parse(html);
+	        FormElement form = ((FormElement) doc.select("form").first());
 
 
-        boolean threw = false;
-        try {
-            Connection con = form.submit();
-        } catch (IllegalArgumentException e) {
-            threw = true;
-            assertEquals("Could not determine a form action URL for submit. Ensure you set a base URI when parsing.",
-                    e.getMessage());
-        }
-        assertTrue(threw);
-    }
+	        boolean threw = false;
+	        try {
+	            Connection con = form.submit();
+	        } catch (IllegalArgumentException e) {
+	            threw = true;
+	            assertEquals("Could not determine a form action URL for submit. Ensure you set a base URI when parsing.",
+	                    e.getMessage());
+	        }
+	        assertTrue(threw);
+	    }
 
-    @Test public void formsAddedAfterParseAreFormElements() {
-        Document doc = Jsoup.parse("<body />");
-        doc.body().html("<form action='http://example.com/search'><input name='q' value='search'>");
-        Element formEl = doc.select("form").first();
-        assertTrue(formEl instanceof FormElement);
+	    @Test public void formsAddedAfterParseAreFormElements() {
+	        Document doc = Jsoup.parse("<body />");
+	        doc.body().html("<form action='http://example.com/search'><input name='q' value='search'>");
+	        Element formEl = doc.select("form").first();
+	        assertTrue(formEl instanceof FormElement);
 
-        FormElement form = (FormElement) formEl;
-        assertEquals(1, form.elements().size());
-    }
+	        FormElement form = (FormElement) formEl;
+	        assertEquals(1, form.elements().size());
+	    }
 
-    @Test public void controlsAddedAfterParseAreLinkedWithForms() {
-        Document doc = Jsoup.parse("<body />");
-        doc.body().html("<form />");
+	    @Test public void controlsAddedAfterParseAreLinkedWithForms() {
+	        Document doc = Jsoup.parse("<body />");
+	        doc.body().html("<form />");
 
-        Element formEl = doc.select("form").first();
-        formEl.append("<input name=foo value=bar>");
+	        Element formEl = doc.select("form").first();
+	        formEl.append("<input name=foo value=bar>");
 
-        assertTrue(formEl instanceof FormElement);
-        FormElement form = (FormElement) formEl;
-        assertEquals(1, form.elements().size());
+	        assertTrue(formEl instanceof FormElement);
+	        FormElement form = (FormElement) formEl;
+	        assertEquals(1, form.elements().size());
 
-        List<Connection.KeyVal> data = form.formData();
-        assertEquals("foo=bar", data.get(0).toString());
-    }
+	        List<Connection.KeyVal> data = form.formData();
+	        assertEquals("foo=bar", data.get(0).toString());
+	    }
 
-    @Test public void usesOnForCheckboxValueIfNoValueSet() {
-        Document doc = Jsoup.parse("<form><input type=checkbox checked name=foo></form>");
-        FormElement form = (FormElement) doc.select("form").first();
-        List<Connection.KeyVal> data = form.formData();
-        assertEquals("on", data.get(0).value());
-        assertEquals("foo", data.get(0).key());
-    }
+	    @Test public void usesOnForCheckboxValueIfNoValueSet() {
+	        Document doc = Jsoup.parse("<form><input type=checkbox checked name=foo></form>");
+	        FormElement form = (FormElement) doc.select("form").first();
+	        List<Connection.KeyVal> data = form.formData();
+	        assertEquals("on", data.get(0).value());
+	        assertEquals("foo", data.get(0).key());
+	    }
 
-    @Test public void adoptedFormsRetainInputs() {
-        // test for https://github.com/jhy/jsoup/issues/249
-        String html = "<html>\n" +
-                "<body>  \n" +
-                "  <table>\n" +
-                "      <form action=\"/hello.php\" method=\"post\">\n" +
-                "      <tr><td>User:</td><td> <input type=\"text\" name=\"user\" /></td></tr>\n" +
-                "      <tr><td>Password:</td><td> <input type=\"password\" name=\"pass\" /></td></tr>\n" +
-                "      <tr><td><input type=\"submit\" name=\"login\" value=\"login\" /></td></tr>\n" +
-                "   </form>\n" +
-                "  </table>\n" +
-                "</body>\n" +
-                "</html>";
-        Document doc = Jsoup.parse(html);
-        FormElement form = (FormElement) doc.select("form").first();
-        List<Connection.KeyVal> data = form.formData();
-        assertEquals(3, data.size());
-        assertEquals("user", data.get(0).key());
-        assertEquals("pass", data.get(1).key());
-        assertEquals("login", data.get(2).key());
-    }
+	    @Test public void adoptedFormsRetainInputs() {
+	        // test for https://github.com/jhy/jsoup/issues/249
+	        String html = "<html>\n" +
+	                "<body>  \n" +
+	                "  <table>\n" +
+	                "      <form action=\"/hello.php\" method=\"post\">\n" +
+	                "      <tr><td>User:</td><td> <input type=\"text\" name=\"user\" /></td></tr>\n" +
+	                "      <tr><td>Password:</td><td> <input type=\"password\" name=\"pass\" /></td></tr>\n" +
+	                "      <tr><td><input type=\"submit\" name=\"login\" value=\"login\" /></td></tr>\n" +
+	                "   </form>\n" +
+	                "  </table>\n" +
+	                "</body>\n" +
+	                "</html>";
+	        Document doc = Jsoup.parse(html);
+	        FormElement form = (FormElement) doc.select("form").first();
+	        List<Connection.KeyVal> data = form.formData();
+	        assertEquals(3, data.size());
+	        assertEquals("user", data.get(0).key());
+	        assertEquals("pass", data.get(1).key());
+	        assertEquals("login", data.get(2).key());
+	    }
+	    
+	    @Test public void TestFormDataNoName(){ 
+	        Document doc = Jsoup.parse("<body />");
+	        doc.body().html("<form />");
+
+	        Element formEl = doc.select("form").first();
+	        formEl.append("<input name value=bar>");
+
+	        assertTrue(formEl instanceof FormElement);
+	        FormElement form = (FormElement) formEl;
+	        assertEquals(1, form.elements().size());
+
+	        List<Connection.KeyVal> data = form.formData();
+	        assertSame(0,data.size());
+	    }
+	    
+	    @Test public void TestFormDataNonSelect(){
+	        String html = "<form><input name='one' value='two'><select name='three'><option value='nop'>" +
+	                "<option value='four'><option value='five'><textarea name=six>seven</textarea>" +
+	                "</form>";
+	        Document doc = Jsoup.parse(html);
+	        FormElement form = (FormElement) doc.select("form").first();
+	        List<Connection.KeyVal> data = form.formData();
+
+	        assertEquals("three=nop", data.get(1).toString());
+	    }
+	    
+	    @Test public void TestFormDataSelectNoOption(){
+	        String html = "<form><input name='one' value='two'><select name='three'>" +
+	                "</form>";
+	        Document doc = Jsoup.parse(html);
+	        FormElement form = (FormElement) doc.select("form").first();
+	        List<Connection.KeyVal> data = form.formData();
+
+	        assertEquals(1, data.size());
+	    }
+
 }
