+++ src/test/java/org/jsoup/nodes/XmlDeclarationTest.java
@@ -1,4 +1,4 @@
-package testoracle;
+package org.jsoup.nodes;
 
 import static org.junit.Assert.*;
