--- src/test/java/testoracle/DocumentTypeTest.java
@@ -1,71 +1,0 @@
-package testoracle;
-
-import static org.junit.Assert.*;
-
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.DocumentType;
-import org.jsoup.parser.Parser;
-import org.junit.Test;
-
-public class DocumentTypeTest {
-	
-	@Test
-    public void constructorValidationOkWithBlankName() {
-        DocumentType fail = new DocumentType("","", "", "");
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void constructorValidationThrowsExceptionOnNulls() {
-        DocumentType fail = new DocumentType("html", null, null, "");
-    }
-
-    @Test
-    public void constructorValidationOkWithBlankPublicAndSystemIds() {
-        DocumentType fail = new DocumentType("html","", "","");
-    }
-
-    @Test public void outerHtmlGeneration() {
-    	DocumentType htmlNull = new DocumentType("", "", "", "");		// Case1 : html is null
-    	assertEquals("<!doctype>", htmlNull.outerHtml());
-    	
-        DocumentType html5 = new DocumentType("html", "", "", "");		// Case2 : html is not null
-        assertEquals("<!doctype html>", html5.outerHtml());
-
-        DocumentType publicDocType = new DocumentType("html", "-//IETF//DTD HTML//", "", "");
-        assertEquals("<!DOCTYPE html PUBLIC \"-//IETF//DTD HTML//\">", publicDocType.outerHtml());
-
-        DocumentType systemDocType = new DocumentType("html", "", "http://www.ibm.com/data/dtd/v11/ibmxhtml1-transitional.dtd", "");
-        assertEquals("<!DOCTYPE html \"http://www.ibm.com/data/dtd/v11/ibmxhtml1-transitional.dtd\">", systemDocType.outerHtml());
-
-        DocumentType combo = new DocumentType("notHtml", "--public", "--system", "");
-        assertEquals("<!DOCTYPE notHtml PUBLIC \"--public\" \"--system\">", combo.outerHtml());
-    }
-
-    @Test public void testRoundTrip() {
-        String base = "<!DOCTYPE html>";
-        assertEquals("<!doctype html>", htmlOutput(base));
-        assertEquals(base, xmlOutput(base));
-
-        String publicDoc = "<!DOCTYPE html PUBLIC \"-//W3C//DTD XHTML 1.0 Transitional//EN\" \"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd\">";
-        assertEquals(publicDoc, htmlOutput(publicDoc));
-        assertEquals(publicDoc, xmlOutput(publicDoc));
-
-        String systemDoc = "<!DOCTYPE html SYSTEM \"exampledtdfile.dtd\">";
-        assertEquals(systemDoc, htmlOutput(systemDoc));
-        assertEquals(systemDoc, xmlOutput(systemDoc));
-
-        String legacyDoc = "<!DOCTYPE html SYSTEM \"about:legacy-compat\">";
-        assertEquals(legacyDoc, htmlOutput(legacyDoc));
-        assertEquals(legacyDoc, xmlOutput(legacyDoc));
-    }
-
-    private String htmlOutput(String in) {
-        DocumentType type = (DocumentType) Jsoup.parse(in).childNode(0);
-        return type.outerHtml();
-    }
-
-    private String xmlOutput(String in) {
-        return Jsoup.parse(in, "", Parser.xmlParser()).childNode(0).outerHtml();
-    }
-
-}
